+++ resources/blocks/e-button/view.blade.php
@@ -23,4 +23,5 @@
   classes="{{ $block_classes ?? '' }}"
   icon="{{ !empty($attributes['button']['showIcon']) && !empty($attributes['button']['icon']['name']) ? $attributes['button']['icon']['name'] : '' }}"
   iconPosition="{{ $attributes['button']['iconPosition'] ?? '' }}"
+  element="{{ $attributes['button']['htmlElement'] ?? 'a' }}"
 />